requirements.txt
Django==3.1.6
beautifulsoup4==4.9.3
bs4==0.0.1
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
lxml==4.6.2
requests==2.25.1
django-extensions==3.1.1
whitenoise==5.2.0
gunicorn==20.0.4
dj-database-url==0.5.0
psycopg2==2.8.6
django_heroku==0.3.1
django_environ==0.4.5
